refactor:LottoUtil의 getRate -> ScoreRankMap의 메소드로 이관
